+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Show project path in tooltip.
+- Show section header above project menu items.
 
 ## [1.0.0] - 2025-05-26